fix(parity): one bad tuning knob no longer discards the whole config.json
ProactiveSkillSubagentsConfig typed minReferences/maxRecommendations as u32, so
serde rejected a non-positive value before the guard could see it — and
load_subagent_extension_config discards the ENTIRE document on any
deserialization error, with only an eprintln. A single `{"minReferences": -1}`
therefore reset maxSubagentDepth, globalConcurrencyLimit, parallel.maxTasks,
every control.* key, chain.dynamicFanout.maxItems, worktreeSetupHook and
maxSubagentDepth to defaults, silently.

Upstream does not: positiveInteger (proactive-skills.ts:32-36) returns undefined
for a value below 1 and the caller falls back to its default, keeping the rest of
the file. cyrup's own guard at discovery/skills.rs:412 was already written for
i64 and handles negatives correctly; the u32 narrowing is what stopped it from
ever running. Widened both fields to i64.

Proven by reverting: with u32, the new test fails `left: 2, right: 5` — the
configured maxSubagentDepth coming back as the default.

Also from the same review:

- extract_fence's retry loop was dead code. The opener begins with the closer's
  own "\n```", so "no closer after this opener" implies "no later opener", and
  the retry can never execute; when a second opener does exist it terminates the
  first one's body on the first pass. Upstream's lazy regex has the same
  reachability, so this is a dead-branch removal, not a behaviour change.
  Replaced with the straight-line equivalent, and the test that claimed to cover
  "skips an unterminated opener" now actually exercises two openers — it had no
  such assertion.

- Three stale citations, each checked against v0.43.0 first: types.ts:880 is
  turnBudget?: TurnBudgetState, not the proactive config (correct targets are
  :1726-1731 for the interface and :1779 for the field), and a v0.34.0
  execution.ts:1054-1056 sat inside an otherwise-v0.43.0 paragraph (:1433-1436).

- An assertion message claimed nothing was "remembered" under a null session id
  when the assertion only shows nothing is surfaced; reworded to what it proves.

Gate: 5941 passed / 0 failed / 8 ignored across 340 suites, run WITH
      ANTHROPIC_API_KEY/OPENAI_API_KEY set to fakes; zero SIGBUS/link failures
      cargo clippy --workspace --all-targets --features test-fixtures: exit 0

+++ crates/cyrup-ext-subagents/src/registration/mod.rs
@@ -135,7 +135,7 @@ pub struct SubagentExtensionConfig {
     pub chain: Option<ExtensionChainConfig>,
     /// Proactive skill-subagent suggestion config — pi
     /// `ExtensionConfig.proactiveSkillSubagents?: ProactiveSkillSubagentsConfig | false`
-    /// (shared/types.ts:840-845/880): an object of tuning knobs, or the literal `false` to disable the
+    /// (shared/types.ts:1726-1731 interface, :1779 field): an object of tuning knobs, or the literal `false` to disable the
     /// feature entirely.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub proactive_skill_subagents: Option<ProactiveSkillSubagents>,
@@ -367,15 +367,26 @@ pub struct ControlConfig {
 pub struct ProactiveSkillSubagentsConfig {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub enabled: Option<bool>,
+    // These are `i64`, not `u32`, so that an out-of-range value REACHES the guard instead of
+    // failing deserialization. `positive_integer` (discovery/skills.rs:412) already filters
+    // `>= 1` and is written for `i64`, matching upstream's `positiveInteger`
+    // (proactive-skills.ts:32-36), which returns `undefined` for a non-positive value and lets
+    // the caller fall back to the default while KEEPING the rest of the file.
+    //
+    // With `u32`, serde rejected `-1` before the guard ever ran, and
+    // `load_subagent_extension_config` (crates/cyrup/src/subagent_config.rs) discards the WHOLE
+    // config.json on any deserialization error — so one bad value silently dropped every other
+    // setting in the file (parallel.maxTasks, control.*, chain.dynamicFanout.maxItems,
+    // worktreeSetupHook, maxSubagentDepth, globalConcurrencyLimit).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_references: Option<u32>,
+    pub min_references: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_recommendations: Option<u32>,
+    pub max_recommendations: Option<i64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub preferred_agent: Option<String>,
 }
 
-/// pi `proactiveSkillSubagents?: ProactiveSkillSubagentsConfig | false` (shared/types.ts:880): either a
+/// pi `proactiveSkillSubagents?: ProactiveSkillSubagentsConfig | false` (shared/types.ts:1779, interface at :1726-1731): either a
 /// tuning-knob object, or the literal `false` to disable the feature entirely. Deserialized
 /// untagged so both a JSON object and a bare `false` parse.
 #[derive(Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
